[#12] Test: ContentSource 유효성 검사 관련 테스트를 PostRequestValidationTest 으로 이동
